[FEATURE] Implemented review exception.

+++ we-re-server/src/reviews/reviews.controller.ts
@@ -21,6 +21,10 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { Review } from 'src/entities/review.entity';
+import {
+  CustomBadTypeRequestException,
+  CustomUnauthorziedException,
+} from 'src/utils/custom_exceptions';
 
 @ApiTags('Reviews')
 @Controller('reviews')
@@ -33,10 +37,15 @@ export class ReviewsController {
     type: [ReadReviewAndWebtoonDto],
   })
   @Get('list/user/:userId')
-  findManyByUserId(
-    @Param('userId') userId: string,
+  async findManyByUserId(
+    @Param('userId') userId: number,
   ): Promise<ReadReviewAndWebtoonDto[]> {
-    return this.reviewsService.findManyByUserId(+userId);
+    try {
+      if (!userId) throw new CustomBadTypeRequestException('userId', userId);
+      return await this.reviewsService.findManyByUserId(userId);
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({ summary: 'create Review' })
@@ -46,7 +55,11 @@ export class ReviewsController {
   })
   @Post()
   createReview(@Body() createReviewDto: CreateReviewDto): Promise<Review> {
-    return this.reviewsService.createReview(createReviewDto);
+    try {
+      return this.reviewsService.createReview(createReviewDto);
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({ summary: 'update Review' })
@@ -60,7 +73,14 @@ export class ReviewsController {
     @Body() updateReviewDto: UpdateReviewDto,
   ): Promise<Review> {
     // param id와 dto 내 id 체크로 자격 여부 판단하는거도 ㄱㅊ할듯
-    return this.reviewsService.updateReview(updateReviewDto);
+    try {
+      if (!id) throw new CustomBadTypeRequestException('id', id);
+      if (id !== updateReviewDto.id)
+        throw new CustomUnauthorziedException(`id is wrong.`);
+      return this.reviewsService.updateReview(updateReviewDto);
+    } catch (error) {
+      throw error;
+    }
   }
 
   @ApiOperation({ summary: 'delete Review' })
@@ -69,6 +89,11 @@ export class ReviewsController {
   @Delete(':id')
   deleteReview(@Param('id') id: number): Promise<void> {
     // 삭제 잘 되었다는 status code 반환~
-    return this.reviewsService.deleteReview(id);
+    try {
+      if (!id) throw new CustomBadTypeRequestException('id', id);
+      return this.reviewsService.deleteReview(id);
+    } catch (error) {
+      throw error;
+    }
   }
 }
